requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest==7.0.1
selenium==4.1.1
webdriver-manager==3.5.3
requests==2.27.1
Faker==13.2.0